1. 소스 코드 패스 잘못 얻어오는 오류 수정 2. "Step Into" 커맨드 (si -> step으로 수정) 3. "Memory" 윈도우 버그 수정

+++ Memory.h
@@ -29,8 +29,8 @@ public:
 	MyGdbFrame *myFrame;
 	
 	int m_column;
-	unsigned int m_start;
-	unsigned int m_count;
+	unsigned long m_start;
+	unsigned long m_count;
 
 private:
 	wxMenu *m_menu;
